requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
cerebras-cloud-sdk>=1.0.0
pydantic>=2.9.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.26.0
sqlalchemy>=2.0.0
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.22
